requirements.txt
openai==1.52.0
httpx==0.27.0
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
